+++ android/app/src/main/kotlin/com/smart_eye/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.smart_eye
+package com.smart_eye
 
 import android.Manifest
 import android.content.Context